% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_doc.R
\name{nba_check_version}
\alias{nba_check_version}
\alias{nba_get_doc}
\title{Check API version}
\usage{
nba_check_version(quiet = TRUE)

nba_get_doc(quiet = FALSE)
}
\arguments{
\item{quiet}{A logical. Should printing be suppressed?}
}
\description{
Retrieve API documentation from \url{https://api.biodiversitydata.nl/v2/reference-doc} and determine whether there was changes since the last package installation.
}
\section{Functions}{
\itemize{
\item \code{nba_get_doc()}: Return a list including the documentation.

}}
\examples{
\dontrun{
 nba_check_version(FALSE)
 nba_get_doc()
}
}
\seealso{
\link{reference_doc}
}
