% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{nbar2}
\alias{nbar2}
\alias{nbar2-package}
\title{nbar2}
\description{
A programmatic interface to the NBA API \url{https://docs.biodiversitydata.nl/endpoints-reference/}.
}
\author{
\strong{Maintainer}: Kevin Cazelles \email{kevin.cazelles@gmail.com} (\href{https://orcid.org/0000-0001-6619-9874}{ORCID})

}
\keyword{internal}
